requirements.txt
streamlit==1.42.2
pillow==11.1.0
google-generativeai==0.8.4
requests==2.32.3
pandas==2.2.3
numpy==2.2.3
